requirements.txt
fusepy==2.0.2
requests==2.7.0
